requirements.txt
pandas
sqlalchemy
python-dotenv
matplotlib
seaborn
